refactor(logic): split built-in node catalogue into per-block modules
The complete catalogue of built-in function blocks lived in one 1136-line
obs/logic/node_types.py, so a one-line change to a single block produced a
diff in the context of every unrelated block and conflicted with every other
branch touching any block.

Each block now lives in its own module under obs/logic/nodes/<category>/,
exporting a single NODE_TYPE. The package name is always identical to the
node's category field, and each category package exports its own NODE_TYPES
tuple. obs/logic/registry.py combines those tuples into the public catalogue
and owns the lookup helpers; registration is explicit rather than scanned, so
a duplicate, misplaced or self-classified node raises on import.

The node definitions are moved verbatim: the assembled catalogue is
definition-identical to before and the serialized GET /api/v1/logic/node-types
payload is byte-identical per node. Only the order of the flat list changes —
it is now strictly grouped by category, following the palette order in
NodePalette.vue. Every consumer resolves by node type, so nothing observes it.

obs/logic/node_types.py stays as a deprecated compatibility facade
re-exporting BUILTIN_NODE_TYPES, NODE_TYPE_REGISTRY, get_node_type and
list_node_types, so branches developed in parallel keep working.

Refs #1109
