+++ obs/logic/nodes/integration/ical.py
@@ -1,0 +1,48 @@
+"""Node definition for the ``ical`` function block (iCalendar)."""
+
+from __future__ import annotations
+
+from obs.logic.models import NodeTypeDef
+from obs.logic.nodes.base import port
+
+NODE_TYPE = NodeTypeDef(
+    type="ical",
+    label="iCalendar",
+    category="integration",
+    description=(
+        "Lädt ein iCal-/ICS-File von einer URL und wertet Termine aus. "
+        "RAW-Ausgang liefert den rohen Kalendertext. "
+        "Pro Filter gibt es 4 Ausgänge: Array (alle Termine), Nächstes Datum, Morgen (Bool), Heute (Bool). "
+        "Filter können auf Summary, Location und/oder Description angewendet werden."
+    ),
+    inputs=[],
+    outputs=[port("raw", "RAW")],  # filter outputs are dynamic — generated by frontend/executor
+    config_schema={
+        "url": {"type": "string", "default": "", "label": "iCal-URL"},
+        "refresh_interval_min": {
+            "type": "number",
+            "default": 60,
+            "label": "Aktualisierungsintervall (Minuten)",
+        },
+        "max_payload_size_mb": {
+            "type": "integer",
+            "default": 2,
+            "min": 1,
+            "max": 50,
+            "label": "Maximale Kalendergrösse (MiB)",
+        },
+        "filter_count": {
+            "type": "integer",
+            "default": 0,
+            "min": 0,
+            "max": 20,
+            "label": "Anzahl Filter",
+        },
+        "filters": {
+            "type": "string",
+            "default": "[]",
+            "label": "Filter (JSON)",
+        },
+    },
+    color="#0369a1",
+)
