+++ obs/logic/nodes/math/avg_multi.py
@@ -1,0 +1,45 @@
+"""Node definition for the ``avg_multi`` function block (Mittelwert)."""
+
+from __future__ import annotations
+
+from obs.logic.models import NodeTypeDef
+from obs.logic.nodes.base import port
+
+NODE_TYPE = NodeTypeDef(
+    type="avg_multi",
+    label="Mittelwert",
+    category="math",
+    description=(
+        "Berechnet den Mittelwert von 2–20 Eingängen (aktuell) und gleitende Mittelwerte "
+        "über Zeitfenster: 1 min, 1 h, 1 Tag, 7/14/30/180/365 Tage. "
+        "Jeder neue Wert wird mit Zeitstempel gespeichert. "
+        "Die Zeitfensterwerte bleiben nach einem Neustart erhalten (konfigurierbar)."
+    ),
+    inputs=[],  # dynamic — generated by frontend based on input_count
+    outputs=[
+        port("avg", "Mittelwert (aktuell)"),
+        port("avg_1m", "Gleit. ∅ 1 min"),
+        port("avg_1h", "Gleit. ∅ 1 Stunde"),
+        port("avg_1d", "Gleit. ∅ 1 Tag"),
+        port("avg_7d", "Gleit. ∅ 7 Tage"),
+        port("avg_14d", "Gleit. ∅ 14 Tage"),
+        port("avg_30d", "Gleit. ∅ 30 Tage"),
+        port("avg_180d", "Gleit. ∅ 180 Tage"),
+        port("avg_365d", "Gleit. ∅ 365 Tage"),
+    ],
+    config_schema={
+        "input_count": {
+            "type": "integer",
+            "default": 2,
+            "min": 2,
+            "max": 20,
+            "label": "Anzahl Eingänge",
+        },
+        "persist_state": {
+            "type": "boolean",
+            "default": True,
+            "label": "Zustand nach Neustart wiederherstellen",
+        },
+    },
+    color="#7c3aed",
+)
