+++ obs/logic/nodes/string/concat.py
@@ -1,0 +1,26 @@
+"""Node definition for the ``string_concat`` function block (String Verketten)."""
+
+from __future__ import annotations
+
+from obs.logic.models import NodeTypeDef
+from obs.logic.nodes.base import port
+
+NODE_TYPE = NodeTypeDef(
+    type="string_concat",
+    label="String Verketten",
+    category="string",
+    description="Verkettet 2–20 Texte zu einem Ergebnis. Jeder Eingang kann dynamisch verbunden oder statisch vorbelegt sein. Ist ein Eingang verbunden, hat er Vorrang vor dem statischen Text.",
+    inputs=[],  # dynamic — generated by frontend based on count
+    outputs=[port("result", "Ergebnis", "string")],
+    config_schema={
+        "count": {
+            "type": "integer",
+            "default": 2,
+            "min": 2,
+            "max": 20,
+            "label": "Anzahl Eingänge",
+        },
+        "separator": {"type": "string", "default": "", "label": "Trennzeichen"},
+    },
+    color="#0891b2",
+)
